# docs/openapi.yaml
...
openapi: 3.0.2
info:
  title: Breeding Insight API
  description: API documentation for Breeding Insight
  termsOfService: http://swagger.io/terms/
  contact:
    email: np398@cornell.edu
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8081/v1
...
paths:
...
  /programs/{programId}/traits:
    get:
      tags:
        - traits
      summary: Get all traits in a given program
      description: Get all traits in a given program
      operationId: getTraits
      parameters:
        - name: programId
          in: path
          description: Id of program to get traits for
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: full
          in: query
          description: True = get detailed properties of traits. False = get only base properties of trait.
          required: false
          explode: false
          schema:
            type: boolean
        - in: query
          description: Field to sort the results by.
          name: sortField
          required: false
          schema:
            type: string
            enum: [name, abbreviations, mainAbbreviation, synonyms, level, status, methodDescription, methodClass, methodFormula, scaleName, scaleClass, scaleDecimalPlaces, scaleLowerLimit, scaleUpperLimit, scaleCategories, createdAt, updatedAt, createdByUserId, createdByUserName, updatedByUserId, updatedByUserName]
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
        - $ref: '#/components/parameters/sortOrder'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/multipleTraitsResponse'
              examples:
                multipleTraitsResponse:
                  $ref: '#/components/examples/arrayOfTraits'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
              example: |-
                ERROR - 2018-10-08T20:15:11Z - Malformed JSON Request Object
                ERROR - 2018-10-08T20:15:11Z - Invalid query parameter
                ERROR - 2018-10-08T20:15:11Z - Required parameter is missing
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - Missing or expired authorization
                token
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - User does not have permission
                to perform this action
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - The requested object DbId is
                not found
    post:
      tags:
        - traits
      summary: Create a list of traits
      description: Create a list of traits in a given program
      operationId: postTraits
      parameters:
        - name: programId
          in: path
          description: Id of program to get traits for
          required: true
          style: simple
          explode: false
          schema:
            type: string
      requestBody:
        description: New trait object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/arrayOfNewTraits'
            example:
              - traitName: "Test Trait"
                traitClass: "morphological"
                attribute: "height"
                defaultValue: "defaultValue"
                mainAbbreviation: "t1"
                abbreviations: ["t1","t2"]
                synonyms: ["stalk height"]
                programObservationLevel:
                  name: "Plant"
                method:
                  methodClass: "Counting"
                  description: "A test method"
                  formula: "a^2+b^2=c^2"
                scale:
                  scaleName: "Test Scale"
                  dataType: "Numerical"
                  decimalPlaces: 3
                  validValueMax: 9999
                  validValueMin: 1
                  categories:
                    - label: "test1"
                      value: "value1"
                    - label: "test2"
                      value: "value2"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/multipleTraitsResponse'
              examples:
                multipleTraitsResponse:
                  $ref: '#/components/examples/arrayOfTraits'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
              example: |-
                ERROR - 2018-10-08T20:15:11Z - Malformed JSON Request Object
                ERROR - 2018-10-08T20:15:11Z - Invalid query parameter
                ERROR - 2018-10-08T20:15:11Z - Required parameter is missing
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - Missing or expired authorization
                token
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - User does not have permission
                to perform this action
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - The requested object DbId is
                not found
        "422":
          description: Multi-error UnprocessableEntity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/validatorErrorResponse"
              example:
                rowErrors:
                  - rowIndex: 1
                    errors:
                      - field: "programObservationLevel"
                        errorMessage: "Missing trait name"
                        httpStatus: "UNPROCESSABLE_ENTITY"
                        httpStatusCode: 422
                      - field: "abbreviations"
                        errorMessage: "Abbreviations already exist"
                        httpStatus: "CONFLICT"
                        httpStatusCode: 409
                      - field: "scale.dataType"
                        errorMessage: "Missing scale data type"
                        httpStatus: "UNPROCESSABLE_ENTITY"
                        httpStatusCode: 422
...
  /programs/{programId}/traits/{traitId}:
    get:
      tags:
        - traits
      summary: Get a single trait by traitId
      description: Gets a single trait with full level of detail
      operationId: getTrait
      parameters:
        - name: programId
          in: path
          description: Id of program to get trait in
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: traitId
          in: path
          description: Id of trait to get trait for
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/singleTraitResponse'
              examples:
                singleTraitResponse:
                  $ref: '#/components/examples/singleTrait'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
              example: |-
                ERROR - 2018-10-08T20:15:11Z - Malformed JSON Request Object
                ERROR - 2018-10-08T20:15:11Z - Invalid query parameter
                ERROR - 2018-10-08T20:15:11Z - Required parameter is missing
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - Missing or expired authorization
                token
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - User does not have permission
                to perform this action
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - The requested object DbId is
                not found
...
components:
  schemas:
    metadata:
      type: object
      properties:
        datafiles:
          type: array
          description: The datafiles key contains a list of file paths, which can
            be relative or complete URLs. These files contain additional information
            related to the returned object and can be retrieved by a subsequent call.
            The empty list should be returned if no additional data files are present.
          items:
            type: string

        pagination:
          $ref: '#/components/schemas/metadata_pagination'
        status:
          type: array
          description: The status field contains a list of informational status messages
            from the server. If no status is reported, an empty list should be returned.
            See Error Reporting for more information.
          items:
            $ref: '#/components/schemas/status'
    status:
      required:
        - message
        - messageType
      type: object
      properties:
        message:
          type: string
          description: A short message concerning the status of this request/response
        messageType:
          type: string
          description: The logging level for the attached message
          enum:
            - DEBUG
            - ERROR
            - WARNING
            - INFO
...
    metadata_pagination:
      type: object
      properties:
        currentPage:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer
      description: The pagination object is applicable only when the payload contains
        a "data" key. It describes the pagination of the data contained in the "data"
        array, as a way to identify which subset of data is being returned. Pages
        are zero indexed, so the first page will be page 0 (zero).
...
    multipleTraitsResponse:
      required:
        - data
      type: object
      properties:
        data:
          type: array
          description: Array of traits
          items:
            $ref: '#/components/schemas/trait'
    singleTraitResponse:
      required:
        - metadata
        - result
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/metadata'
        result:
          $ref: '#/components/schemas/trait'
    trait:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: System id of the trait
        traitName:
          type: string
          description: Trait display name
        active:
          type: boolean
          description: Trait active status
        createdAt:
          type: string
          format: datetime
        updatedAt:
          type: string
          format: datetime
        createdByUser:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        updatedByUser:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        traitClass:
          type: string
          example: "morphological"
        attribute:
          type: string
          description: Attribute the trait is describing
          example: height
        defaultValue:
          type: string
          description: Default value of the trait
        mainAbbreviation:
          type: string
        abbreviations:
          type: array
          items:
            type: string
        synonyms:
          type: array
          items:
            type: string
        programObservationLevel:
          type: object
          properties:
            id:
              type: string
              format: uuid
              description: System id of the observation level
            name:
              type: string
              description: Name of the observation level
        method:
          type: object
          properties:
            methodClass:
              type: string
            description:
              type: string
            formula:
              type: string
        scale:
          type: object
          properties:
            scaleName:
              type: string
            dataType:
              type: string
            decimalPlaces:
              type: integer
            categories:
              type: array
              items:
                type: object
                properties:
                  label:
                    type: string
                  value:
                    type: string
            validValueMax:
              type: integer
            validValueMin:
              type: integer
...
    arrayOfNewTraits:
      type: array
      items:
        $ref: '#/components/schemas/newTrait'
    newTrait:
      type: object
      properties:
        traitName:
          type: string
          description: Trait display name
        traitClass:
          type: string
          example: "morphological"
        attribute:
          type: string
          description: Attribute the trait is describing
          example: height
        defaultValue:
          type: string
          description: Default value of the trait
        mainAbbreviation:
          type: string
        abbreviations:
          type: array
          items:
            type: string
        synonyms:
          type: array
          items:
            type: string
        programObservationLevel:
          type: object
          properties:
            name:
              type: string
              description: Name of the observation level
        method:
          type: object
          properties:
            methodClass:
              type: string
            description:
              type: string
              format: UUID
            formula:
              type: string
              format: UUID
        scale:
          type: object
          properties:
            scaleName:
              type: string
              format: datetime
            dataType:
              type: string
              format: datetime
            decimalPlaces:
              type: integer
            categories:
              type: array
              items:
                type: object
                properties:
                  label:
                    type: string
                  value:
                    type: string
            validValueMax:
              type: integer
            validValueMin:
              type: integer
    validatorErrorResponse:
      type: object
      properties:
        rowErrors:
          type: array
          items:
            type: object
            properties:
              rowIndex:
                type: integer
                description: "Row number in the uploaded file for the trait with errors."
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      description: "Field name that is the source of the error"
                    errorMessage:
                      type: string
                      description: "A description fo the error"
                    httpStatus:
                      type: string
                      description: "Text representation fo the error code"
                    httpStatusCode:
                      type: integer
                      description: "Http error code"
...
  securitySchemes:
    orcid_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://sandbox.orcid.org/oauth/authorize
          scopes:
            write:bi: modify
            read:bi: read
    api_key:
      type: apiKey
      name: api_key
      in: header
  examples:
...
    arrayOfTraits:
      value:
        metadata:
          datafiles: []
          pagination:
            currentPage: 1
            pageSize: 1
            totalCount: 1
            totalPages: 1
          status: []
        result:
          data:
            [
            {
              id: "00385063-01c8-4afc-a684-3c84b90b2de9",
              traitName: "Test Trait",
              active: true,
              createdAt: "2020-06-04T11:05:48-04:00",
              updatedAt: "2020-06-04T11:05:48-04:00",
              traitClass: "morphological",
              attribute: "height",
              defaultValue: "defaultValue",
              entity: "stalk",
              mainAbbreviation: "t1",
              abbreviations: ["t1","t2"],
              synonyms: ["stalk height"],
              programObservationLevel:
                {
                  id: "71421c6a-19d4-4d82-858b-e3f70a9c7a02",
                  name: "Plant",
                  active: true,
                  createdAt: "2020-06-04T11:05:48-04:00",
                  updatedAt: "2020-06-04T11:05:48-04:00",
                },
              method:
                {
                  methodClass: "Counting"
                },
              scale:
                {
                  scaleName: "Test Scale",
                  dataType: "TEXT",
                  decimalPlaces: 3,
                  categories:
                    [
                    {
                      label: "test1",
                      value: "value1"
                    },
                    {
                      label: "test2",
                      value: "value2"
                    }
                    ]
                },
              programOntology:
                {
                  id: "a9712526-85ca-4a47-afe6-974ce4abf6a9",
                  programId: "c9caac57-3dcb-442b-b3e9-e01281fbf8bd",
                  createdAt: "2020-06-04T11:05:48-04:00",
                  updatedAt: "2020-06-04T11:05:48-04:00"
                },
              createdByUser:
                {
                  name: "Tester",
                  active: true
                },
              updatedByUser:
                {
                  name: "Tester",
                  active: true
                }
            }
            ]
    singleTrait:
      value:
        metadata:
          datafiles: []
          pagination:
            currentPage: 1
            pageSize: 1
            totalCount: 1
            totalPages: 1
          status: []
        result:
          {
            id: "00385063-01c8-4afc-a684-3c84b90b2de9",
            traitName: "Test Trait",
            active: true,
            createdAt: "2020-06-04T11:05:48-04:00",
            updatedAt: "2020-06-04T11:05:48-04:00",
            traitClass: "morphological",
            attribute: "height",
            defaultValue: "defaultValue",
            entity: "stalk",
            mainAbbreviation: "t1",
            abbreviations: ["t1","t2"],
            synonyms: ["stalk height"],
            programObservationLevel:
              {
                id: "71421c6a-19d4-4d82-858b-e3f70a9c7a02",
                name: "Plant",
                active: true,
                createdAt: "2020-06-04T11:05:48-04:00",
                updatedAt: "2020-06-04T11:05:48-04:00",
              },
            method:
              {
                methodClass: "Counting"
              },
            scale:
              {
                scaleName: "Test Scale",
                dataType: "TEXT",
                decimalPlaces: 3,
                categories:
                  [
                  {
                    label: "test1",
                    value: "value1"
                  },
                  {
                    label: "test2",
                    value: "value2"
                  }
                  ]
              },
            programOntology:
              {
                id: "a9712526-85ca-4a47-afe6-974ce4abf6a9",
                programId: "c9caac57-3dcb-442b-b3e9-e01281fbf8bd",
                createdAt: "2020-06-04T11:05:48-04:00",
                updatedAt: "2020-06-04T11:05:48-04:00"
              },
            createdByUser:
              {
                name: "Tester",
                active: true
              },
            updatedByUser:
              {
                name: "Tester",
                active: true
              }
          }
...
  parameters:
    page:
      description: 'Used to request a specific page of data to be returned.
        The page indexing starts at 1 (the first page is ''page''= 1). Default is `1`.'
      example: 1
      in: query
      name: page
      required: false
      schema:
        type: integer
    pageSize:
      description: The size of the pages to be returned. Default is to show all results.
      example: 1000
      in: query
      name: pageSize
      required: false
      schema:
        type: integer
    sortOrder:
      description: Order in which to sort the results. Default is ascending order.
      example: ASC
      in: query
      name: sortOrder
      required: false
      schema:
        type: string